requirements.txt
discord.py
asyncio
requests
urllib
random
